[optimizer] rename files
